Rewrote config
Former-commit-id: 9a66048dfe84d929c911ff7e0d42a9d7f86f1ee5

+++ Intersect Library/Intersect Library/Options.cs
@@ -9,7 +9,7 @@ namespace Intersect
         //Game Settings
         public static string Language = "English";
         public static string GameName = "Intersect";
-        public static int ServerPort = 4500;
+        public static ushort ServerPort = 4500;
 
         //Maxes
         public static int MaxNpcDrops = 10;
